backend/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
google-auth==2.22.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.88.0
requests==2.31.0
sqlalchemy==2.2.4
openai==1.0.0
pytz==2025.3
